requirements.txt
beautifulsoup4==4.12.3
faiss==1.9.0
faiss-gpu==1.9.0
huggingface_hub==0.29.2
numpy==1.26.3
python-dotenv==1.0.1
requests==2.32.3
sentence-transformers==3.4.1
streamlit==1.43.2
torch==2.6.0